requirements.txt
# Library to call various geocoders
geopy>=2.0.0

# Optional import for geopy to use HTTP keepalive
requests>=2.23.0

# Library to communicate easily with Exiftool
PyExifTool==0.5.4